nah b+ dokumentum

+++ src/app/dokumentum/dokumentum-egy/dokumentum-egy.component.ts
@@ -13,9 +13,8 @@ import {EgyMode} from '../../enums/egymode';
 export class DokumentumEgyComponent implements OnDestroy {
   egymode = EgyMode.Reszletek;
   dokumentumservice: DokumentumService;
-  ri = -1;
 
-  @Output() torlesutan = new EventEmitter<void>();
+  @Output() eventTorlesutan = new EventEmitter<void>();
 
   private _eppFrissit = false;
   get eppFrissit(): boolean {
@@ -32,12 +31,13 @@ export class DokumentumEgyComponent implements OnDestroy {
     this.dokumentumservice = dokumentumservice;
   }
 
-  reszletek() {
+  doReszletek() {
     this.egymode = EgyMode.Reszletek;
   }
-  torles() {
+  doTorles() {
     this.egymode = EgyMode.Torles;
   }
+
   letoltes() {
     this.eppFrissit = true;
     this.dokumentumservice.Kimentes()
@@ -61,32 +61,29 @@ export class DokumentumEgyComponent implements OnDestroy {
       });
   }
 
-  TorlesOk() {
-    this.eppFrissit = true;
-    this.dokumentumservice.Delete(this.dokumentumservice.Dto[this.dokumentumservice.DtoSelectedIndex])
-      .then(res => {
-        if (res.Error != null) {
-          throw res.Error;
-        }
+  onTorles(ok: boolean) {
+    if (ok) {
+      this.eppFrissit = true;
 
-        this.dokumentumservice.Dto.splice(this.dokumentumservice.DtoSelectedIndex, 1);
-        this.dokumentumservice.DtoSelectedIndex = -1;
+      this.dokumentumservice.Delete(this.dokumentumservice.Dto[this.dokumentumservice.DtoSelectedIndex])
+        .then(res => {
+          if (res.Error != null) {
+            throw res.Error;
+          }
 
-        this.eppFrissit = false;
-        this.torlesutan.emit();
-      })
-      .catch(err => {
-        this.eppFrissit = false;
-        this._errorservice.Error = err;
-      });
-  }
+          this.dokumentumservice.Dto.splice(this.dokumentumservice.DtoSelectedIndex, 1);
+          this.dokumentumservice.DtoSelectedIndex = -1;
 
-  TorlesCancel() {
-    this.egymode = EgyMode.Reszletek;
-  }
-
-  EgyReszletek() {
-    this.egymode = EgyMode.Reszletek;
+          this.eppFrissit = false;
+          this.eventTorlesutan.emit();
+        })
+        .catch(err => {
+          this.eppFrissit = false;
+          this._errorservice.Error = err;
+        });
+    } else {
+      this.egymode = EgyMode.Reszletek;
+    }
   }
 
   ngOnDestroy() {
